add ints.h

+++ src/queue.cuh
@@ -1,15 +1,15 @@
-#include <stdint.h>
 #include <stdio.h>
+#include "ints.h"
 
 typedef struct Queue
 {
-    uint8_t len;
-    uint8_t r_ptr;
-    uint8_t w_ptr;
-    uint8_t buf[16];
+    u8 len;
+    u8 r_ptr;
+    u8 w_ptr;
+    i16 buf[16];
 } Queue_t;
 
-__device__ int queue_read(Queue_t *q, uint8_t *n)
+__device__ u8 queue_read(Queue_t *q, i16 *n)
 {
     if (q->len == 0)
     {
@@ -24,7 +24,7 @@ __device__ int queue_read(Queue_t *q, uint8_t *n)
     return 0;
 }
 
-__device__ int queue_write(Queue_t *q, uint8_t n)
+__device__ u8 queue_write(Queue_t *q, i16 n)
 {
     if (q->len == 16)
     {
